pbt-equijoin/requirements.txt
hypothesis>=6.0
psycopg2-binary>=2.9
pytest>=7.0
